+++ lib/tkt.c
@@ -91,7 +91,7 @@ shishi_tkt_build (Shishi_tkt * tkt, Shishi_key * key)
  * Return value: Returns client principal of ticket.
  **/
 int
-shishi_tkt_client (Shishi_tkt * tkt, char *client, int *clientlen)
+shishi_tkt_client (Shishi_tkt * tkt, char *client, size_t *clientlen)
 {
   return shishi_principal_name_get (tkt->handle, tkt->kdcrep,
 				    "cname", client, clientlen);
@@ -101,7 +101,7 @@ int
 shishi_tkt_client_p (Shishi_tkt * tkt, const char *client)
 {
   char *buf;
-  int buflen;
+  size_t buflen;
   int res;
 
   buflen = strlen (client) + 1;
@@ -127,7 +127,8 @@ shishi_tkt_client_p (Shishi_tkt * tkt, const char *client)
 }
 
 int
-shishi_tkt_cnamerealm (Shishi_tkt * tkt, char *cnamerealm, int *cnamerealmlen)
+shishi_tkt_cnamerealm (Shishi_tkt * tkt,
+		       char *cnamerealm, size_t *cnamerealmlen)
 {
   return shishi_principal_name_realm_get (tkt->handle,
 					  tkt->kdcrep, "cname",
@@ -573,7 +574,7 @@ shishi_tkt_realm (Shishi_tkt * tkt, char **realm, size_t *realmlen)
 }
 
 int
-shishi_tkt_server (Shishi_tkt * tkt, char *server, int *serverlen)
+shishi_tkt_server (Shishi_tkt * tkt, char *server, size_t *serverlen)
 {
   return shishi_ticket_sname_get (tkt->handle, tkt->ticket,
 				  server, serverlen);
@@ -583,7 +584,7 @@ int
 shishi_tkt_server_p (Shishi_tkt * tkt, const char *server)
 {
   char *buf;
-  int buflen;
+  size_t buflen;
   int res;
 
   buflen = strlen (server) + 1;
@@ -610,7 +611,7 @@ shishi_tkt_server_p (Shishi_tkt * tkt, const char *server)
 
 int
 shishi_tkt_server_realm (Shishi_tkt * tkt,
-			 char *serverrealm, int *serverrealmlen)
+			 char *serverrealm, size_t *serverrealmlen)
 {
   return shishi_ticket_snamerealm_get (tkt->handle, tkt->ticket,
 				       serverrealm, serverrealmlen);
@@ -639,7 +640,7 @@ shishi_tkt_keytype_p (Shishi_tkt * tkt, int32_t etype)
 
 int
 shishi_tkt_lastreq (Shishi_tkt * tkt,
-		    char *lrtime, int *lrtimelen, int lrtype)
+		    char *lrtime, size_t *lrtimelen, int lrtype)
 {
   char *format;
   int tmplrtype;
@@ -728,7 +729,7 @@ shishi_tkt_lastreq_pretty_print (Shishi_tkt * tkt, FILE * fh)
 }
 
 int
-shishi_tkt_authtime (Shishi_tkt * tkt, char *authtime, int *authtimelen)
+shishi_tkt_authtime (Shishi_tkt * tkt, char *authtime, size_t *authtimelen)
 {
   return shishi_asn1_read (tkt->handle, tkt->enckdcreppart, "authtime",
 			   authtime, authtimelen);
@@ -755,7 +756,7 @@ shishi_tkt_authctime (Shishi_tkt * tkt)
 }
 
 int
-shishi_tkt_starttime (Shishi_tkt * tkt, char *starttime, int *starttimelen)
+shishi_tkt_starttime (Shishi_tkt * tkt, char *starttime, size_t *starttimelen)
 {
   return shishi_asn1_read_optional (tkt->handle, tkt->enckdcreppart,
 				    "starttime", starttime, starttimelen);
@@ -782,7 +783,7 @@ shishi_tkt_startctime (Shishi_tkt * tkt)
 }
 
 int
-shishi_tkt_endtime (Shishi_tkt * tkt, char *endtime, int *endtimelen)
+shishi_tkt_endtime (Shishi_tkt * tkt, char *endtime, size_t *endtimelen)
 {
   return shishi_asn1_read (tkt->handle, tkt->enckdcreppart, "endtime",
 			   endtime, endtimelen);
@@ -809,7 +810,7 @@ shishi_tkt_endctime (Shishi_tkt * tkt)
 }
 
 int
-shishi_tkt_renew_till (Shishi_tkt * tkt, char *renewtill, int *renewtilllen)
+shishi_tkt_renew_till (Shishi_tkt * tkt, char *renewtill, size_t *renewtilllen)
 {
   return shishi_asn1_read_optional (tkt->handle, tkt->enckdcreppart,
 				    "renew-till", renewtill, renewtilllen);
